Update NPM packages, update webpacker and configs.
This includes a ton of configuration tweaks to get the app setup for the tools I want to use.

I decided to lean on StimulusJS vs React since I like how its footprint is a lot leaner. I'm a fan of how Stimulus keeps things super light and how it has a great synergy with Rails UJS.

I had to update Webpacker to version 4.2.2 since I wanted to lean on TailwindCSS for my styling. So this also includes a ton of configuration changes to co-incide with the webpacker upgrade.