Remove sample React App, update Application.js, add styling.
This removes the sample React App that was generated with the Rails new command.

This adds some basic styling to the application and a generated controller for Stimulus.

Some of this styling was written adhoc for this application, and some of it was ported over from a side project I've been working on.