Setup basic routes and seed data.
This sets up the basic routes we'll be needing to modify our application resources.

This also sets up some seed data for us so we can have a reasonable amount of sample data to play with in the app.

Scotch is indeed the essence of life. Ron Swanson is a wise man. :beers: