Add Post Controllers.
This adds a couple of controllers that form the bulk of the feature logic for Posts.

1) The Posts Controllers are relatively straight forward.

Posts#index hooks into the PG Search gem, with Kaminari, to offer search + pagination for results.

The create and destroy action uses both old school Rails full re-directs and some Stimulus/Rails UJS Ajax goodness when creating a new Post. These patterns could be leveraged to add editing of existing Posts to the application as well.