Add Comments Controllers, both Rails and Stimulus! Add Index.js for imports.
The Comments Controllers have a lot of logic in them due to the fact that Comment creation, deletion, and editing are all AJAXified to give the application a nice SPA type feel, while leveraging Rails form builders and established paradigms.

The StimulusJS Comments Controller could use a tiny bit of refactoring, but due to time constraints I'm happy with where it is at.